+++ examples/06 - input/script.py
@@ -2,11 +2,10 @@ import os
 import sys
 
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), '..'))
-from common import WINDOW_SIZE, use_shared_assets
+from common import WINDOW_SIZE
 
 import glue
 
-use_shared_assets()
 glue.init('web')                     # Give folder containing web files
 
 @glue.expose                         # Expose this function to Javascript