+++ examples/10 - custom_app_routes/custom_app.py
@@ -2,7 +2,7 @@ import os
 import sys
 
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), '..'))
-from common import WINDOW_SIZE, use_shared_assets
+from common import WINDOW_SIZE
 
 import glue
 import bottle
@@ -14,7 +14,6 @@ app = bottle.Bottle()
 def custom_route():
     return 'Hello, World!'
 
-use_shared_assets(app)
 glue.init('web')
 
 # need to manually add glue routes if we are wrapping our Bottle instance with middleware